New controller for bookswap email functionality
